Initial commit extending Chart.js bar chart
Bar charts now require an array for the fillColor value in the dataset
objects.

The ultimate goal is to allow either single values for fillColor or an
array of values for each datapoint. Thus you can show multivariate
datasets on a single bar chart with a single set of bars. I.e. you have
one set of labels for the graph, for each dataset in the graph you can
associate unique styling attributes for each bar for that dataset.
Currently you can have one set of attributes for a whole dataset.