Converted parseCrimeData to use _.js
I cleaned up a lot of the code with the help of _.omit, as well as
re-evaluating what I was looping through the objects for. So where I
used to loop through possibleCrimes to pull values/objects out of, say,
crimeRanges, now I simply use _.each to loop through the object itself.
I could have done this before using vanilla js, most likely, but this
came back to me by refactoring the code.

I also renamed metaCrimes to crimeFilter, as I included 'NEIGHBORHOOD'
as well, which isn't even a meta crime. metaCrimes was a less accurate
name as well.