Commented out test code
While I wait for the PR for Chart.js I've commented out code I used to
test my 'arrays for styling' addition.

I also moved the fill colors assignment to the top of the code, as it
never changes once set (the color scheme for the crimes is set once and
once only).