Major refactoring
I didn't like how much code was inside createMap, as it wasn't just
about the map any more, and it was also hard to assess what was about
the map, the info box, the crimeselecter and the topo layer. So I pulled
those out into seperate functions for easier handling.