Got the crimes to be listed one/line
Now the crimes in the dropdown are listed one per line, whereas before
it lumped a couple at the bottom.

Next up is to figure out why the buttons have the "radio button" circle
and get rid of it. Also, change the styling of the buttons so they fit
the background of the dropdown.